Create README.md
I received a request to make a whatsapp bot that would fallow all the Whatsapp requirement and that wouldn't be banned. The main objective were:
1. Add new contacts to a CSV and CVF file that are later imported to a phone
2. While importing a new number send a welcome message

Other functions were added when they were needed. This is a BEGINNER friendly repository as I am using it to learn GIT. Any contribution is welcome.